Get the right number of rating. Next step: to add this number in database

+++ app/src/main/java/com/buzuriu/dogapp/models/ReviewObj.kt
@@ -2,11 +2,11 @@ package com.buzuriu.dogapp.models
 
 class ReviewObj {
     var userIdThatLeftReview: String? = null
-    var numberOfStars : Int? = null
+    var numberOfStars : Float? = null
 
     constructor()
 
-    constructor(userIdThatLeftReview: String, numberOfStars: Int) {
+    constructor(userIdThatLeftReview: String, numberOfStars: Float) {
         this.userIdThatLeftReview = userIdThatLeftReview
         this.numberOfStars = numberOfStars
     }
